--- frontend/src/App.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Search from './components/Search'
-import './App.css';
-
-const App = () => {
-  return (
-    <div className="App">
-      <Search />
-    </div>
-  );
-}
-
-export default App;
