+++ frontend/src/components/Gifs.css
@@ -2,5 +2,5 @@
   grid-column-start: 2;
   grid-column-end: 2;
   grid-row-start: 3;
-  transition: all 1s ease 1s
+  transition: all 1s ease 1s;
 }
