+++ frontend/src/components/Gifs.jsx
@@ -4,18 +4,12 @@ import Masonry from 'react-masonry-component';
 import './Gifs.css'
 
 const Gifs = props => {
-  const masonryOptions = {
-    transitionDuration: 1
-  };
-
   return (
     <div className="gifs">
       <Masonry
         className={'my-gallery-class'} // default ''
-        options={masonryOptions} // default {}
         disableImagesLoaded={false} // default false
         updateOnEachImageLoad={false} // default false and works only if disableImagesLoaded is false
-        options={masonryOptions}
       >
         {props.gifs.map(function(gif, i){
           return(
