+++ frontend/src/components/Search.jsx
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
+import { Link } from "react-router-dom";
 import Gifs from './Gifs'
 import SearchBar from './SearchBar'
 import axios from 'axios'
@@ -6,22 +7,8 @@ import './Search.css'
 
 const Search = () => {
   const [gifs, setGifs] = useState([])
-  const [favorites, setFavorites] = useState([])
   const [search, setSearch] = useState('')
   const [validSearch, setValidSearch] = useState(false)
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const result = await axios({
-        method: 'GET',
-        url: 'http://localhost:3010/api/favorites'
-      })
-      .then(response => {
-        setFavorites(response.data)
-      });
-    }
-    fetchData()
-  }, []);
 
   const handleSubmit = e => {
     e.preventDefault()
@@ -30,7 +17,14 @@ const Search = () => {
       url: `http://localhost:3010/api/gifs/search?search=${search}`
     })
     .then(response => {
-      setGifs(response.data.data)
+      let searchGifs = response.data.data.map((gif) => {
+        return {
+          id: gif.id,
+          title: gif.title,
+          url: gif.images.downsized.url
+        }
+      })
+      setGifs(searchGifs)
     })
     .catch(error => {
       console.log(error)
@@ -54,10 +48,11 @@ const Search = () => {
 
   return (
     <div className="search">
+      <Link to="/favorites" className="link">Favorites</Link>
       <SearchBar 
-        handleSubmit={handleSubmit}
-        handleInput={handleInput}
-        validSearch={validSearch}
+      handleSubmit={handleSubmit}
+      handleInput={handleInput}
+      validSearch={validSearch}
       />
       <Gifs gifs={gifs} />
     </div>
