+++ frontend/src/components/SearchBar.css
@@ -34,7 +34,3 @@ button {
   border: 1px solid currentColor;
   background: transparent;
 }
-
-.primary {
-  border-color: rgb(148, 255, 159);
-}
